+++ pisek/cms/dataset.py
@@ -21,7 +21,7 @@ import datetime
 
 from pisek.cms.testcase import create_testcase, get_testcases
 from pisek.env.env import Env
-from pisek.env.task_config import JudgeType, TaskConfig, TaskType
+from pisek.config.task_config import JudgeType, TaskConfig, TaskType
 from pisek.utils.paths import TaskPath
