+++ pisek/config/global-defaults
