+++ pisek/config/task_config.py
@@ -34,7 +34,7 @@ from typing import Optional, Any, Annotated, Union
 from pisek.utils.text import tab
 from pisek.utils.text import eprint, colored, warn
 from pisek.env.base_env import BaseEnv
-from pisek.env.config_hierarchy import ConfigValue, TaskConfigError, ConfigHierarchy
+from pisek.config.config_hierarchy import ConfigValue, TaskConfigError, ConfigHierarchy
 from pisek.env.context import init_context
 from pisek.jobs.parts.solution_result import SUBTASK_SPEC
