+++ pisek/jobs/parts/tools.py
@@ -22,7 +22,7 @@ import subprocess
 from pisek.jobs.jobs import State, Job, PipelineItemFailure
 from pisek.env.env import Env
 from pisek.utils.paths import TaskPath
-from pisek.env.task_config import ProgramType
+from pisek.config.task_config import ProgramType
 from pisek.jobs.parts.task_job import TaskJob, TaskJobManager
 from pisek.jobs.parts.program import ProgramsJob
