+++ pisek/visualize.py
@@ -18,8 +18,8 @@ from typing import Optional
 
 from pisek.utils.text import pad, tab, colored, eprint
 from pisek.utils.terminal import terminal_width
-from pisek.env.task_config import load_config, TaskConfig, FailMode
-from pisek.env.select_solutions import expand_solutions, UnknownSolutions
+from pisek.config.task_config import load_config, TaskConfig, FailMode
+from pisek.config.select_solutions import expand_solutions, UnknownSolutions
 from pisek.jobs.parts.solution_result import Verdict
 from pisek.jobs.parts.verdicts_eval import evaluate_verdicts
